ui: Hoist lezer grammar file parsing into the bundler using the lezer rollup plugin
