+++ ui/config/grammar_mock.js
@@ -1,0 +1,18 @@
+// Copyright (C) 2026 The Android Open Source Project
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Stub for *.grammar imports under jest. Vite compiles these via
+// @lezer/generator in production; tests don't exercise the parser, so a
+// no-op `parser` object is enough to satisfy the import.
+module.exports = {parser: {}};
