+++ ui/package.json
@@ -51,6 +51,7 @@
   "devDependencies": {
     "@eslint/eslintrc": "^3.1.0",
     "@eslint/js": "^9.6.0",
+    "@lezer/generator": "^1.8.0",
     "@playwright/test": "^1.58.2",
     "@rollup/plugin-commonjs": "^28.0.2",
     "@rollup/plugin-node-resolve": "^16.0.0",
