--- ui/src/base/perfetto_sql_lang/gen/perfetto_sql.grammar.terms.js
@@ -1,12 +1,0 @@
-// This file was generated by lezer-generator. You probably shouldn't edit it.
-export const
-  LineComment = 1,
-  Program = 2,
-  String = 3,
-  Number = 4,
-  Boolean = 5,
-  Identifier = 6,
-  Null = 7,
-  Keyword = 8,
-  Function = 9,
-  MacroVariable = 10
