+++ ui/src/base/perfetto_sql_lang/language.ts
@@ -21,43 +21,53 @@ import {
   indentNodeProp,
 } from '@codemirror/language';
 import {styleTags, tags as t} from '@lezer/highlight';
-import {parser} from './gen/perfetto_sql.grammar'; // Generated by Lezer
+import {parser} from './perfetto_sql.grammar'; // Compiled on import by Vite's lezer plugin.
 import type {TreeCursor} from '@lezer/common';
 
-export const language = LRLanguage.define({
-  parser: parser.configure({
-    props: [
-      indentNodeProp.add({
-        Application: delimitedIndent({closing: ')', align: false}),
-      }),
-      foldNodeProp.add({
-        Application: foldInside,
-      }),
-      styleTags({
-        'Keyword': t.keyword,
-        'Boolean': t.bool,
-        'Null': t.null,
-        'String': t.string,
-        'Number': t.number,
-        'LineComment': t.lineComment,
-        'Macro': t.macroName,
-        'MacroVariable': t.variableName,
-        'Function': t.function(t.variableName),
-        '( )': t.paren,
-      }),
-    ],
-  }),
-  languageData: {
-    commentTokens: {line: '--'},
-  },
-});
+let cache: LRLanguage | undefined;
+
+function createLanguage() {
+  // Laziy initialize the language object. The reason we memoize it is becuase
+  // the generated language is always the same so it's always safe to reuse. The
+  // reason we don't create it once at the top level is because this file is
+  // imported in a lot of tests but never used so we want to:
+  // 1. Avoid the overhead of creating the language in those cases.
+  // 2. Be able to stub out the import.
+  return (cache ??= LRLanguage.define({
+    parser: parser.configure({
+      props: [
+        indentNodeProp.add({
+          Application: delimitedIndent({closing: ')', align: false}),
+        }),
+        foldNodeProp.add({
+          Application: foldInside,
+        }),
+        styleTags({
+          'Keyword': t.keyword,
+          'Boolean': t.bool,
+          'Null': t.null,
+          'String': t.string,
+          'Number': t.number,
+          'LineComment': t.lineComment,
+          'Macro': t.macroName,
+          'MacroVariable': t.variableName,
+          'Function': t.function(t.variableName),
+          '( )': t.paren,
+        }),
+      ],
+    }),
+    languageData: {
+      commentTokens: {line: '--'},
+    },
+  }));
+}
 
 export function perfettoSql() {
-  return new LanguageSupport(language);
+  return new LanguageSupport(createLanguage());
 }
 
 export function parseAndPrintTree(code: string) {
-  const tree = language.parser.parse(code);
+  const tree = createLanguage().parser.parse(code);
   const cursor = tree.cursor();
   printCST(cursor, code);
 }
