+++ ui/src/core_plugins/dev.perfetto.CoreCommands/legacy_macros_schema.ts
@@ -1,0 +1,22 @@
+// Copyright (C) 2026 The Android Open Source Project
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import {commandInvocationSchema} from '../../core/command_manager';
+import z from 'zod';
+
+// Legacy macro schema (dictionary format) - deprecated, kept for migration
+export const legacyMacrosConfigSchema = z.record(
+  z.string(), // key: macro name
+  z.array(commandInvocationSchema).readonly(),
+);
